Remove unused `DataOutputWritable`
